Решение уравения теплопроводности 3D для куба  10на10, 4 усточника в центре куба

+++ CoRSaD/VoxelLibrary/VoxelLibrary/ColorVoxel.cs
@@ -20,7 +20,7 @@ namespace VoxelLibrary
             green = 0;
             blue = 0;
         }
-        public ColorVoxel(float sMin, float sMax)//диапазон напряяжений
+        public ColorVoxel(double sMin, double sMax)//диапазон напряяжений
         {
             red = 0;
             green = 1;
